requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.0
anthropic==0.7.0
chromadb==0.4.15
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7
pydantic-settings==2.9.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
jsonschema==4.20.0
pyyaml==6.0.1
requests==2.31.0
numpy==1.24.3
pandas==2.1.4
loguru==0.7.2
